-Добавил тесты для класса ApiController -Добавил конструктор с userId в ApiUser -Изменил возвращаемый тип метода userCountByLogin в UserRepo

+++ src/main/java/com/nadir/vk_internship/repository/UserRepo.java
@@ -12,5 +12,5 @@ public interface UserRepo extends JpaRepository<ApiUser, Integer> {
     ApiUser authenticate(String login, String pswd);
 
     @Query(value = "SELECT COUNT(1) FROM users WHERE login = ?1",nativeQuery = true)
-    long userCountByLogin(String login);
+    short userCountByLogin(String login);
 }
